user/pyproject.toml
[project]
name = "user"
version = "0.1.0"
description = ""
authors = [
    {name = "albandvd",email = "alban.david06@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.4,<47.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "user", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
